+++ CraterMine/core/src/com/img/cratermine/Entities/PlayerStats.java
@@ -1,0 +1,15 @@
+package com.img.cratermine.Entities;
+
+public class PlayerStats {
+	private int attack;
+	private int gold;
+	private int minerals;
+	
+	void setAtributes(int attack, int gold, int minerals){
+		this.attack = attack;
+		this.gold = gold;
+		this.minerals = minerals;
+	}
+
+}
+
